my resources lied to me :(
when i was doing my javascript, i looked up how i should do and ask ChatGPT to TEACH me how to do it so I can write it and actually understand. HOWEVER it decided to have me write connect by ID and the DOMconnected terms.
Luckly, reviewing at a previous example, I managed to go in and fix it, as well as making it still work.!